fix(symthaea): lint cleanups + hardened assertions + security tests
- Remove redundant `as f32`/`as f64` casts in cycle.rs and symthaea.rs
- Upgrade debug_assert to assert with descriptive messages in ContinuousHV add/subtract
- Add npy shape overflow and too-large-array security tests in neural_bridge
- Add invalid contribution filtering test in hybrid_bft

+++ src/cognitive_loop/cycle.rs
@@ -2522,7 +2522,7 @@ impl CognitiveLoopService {
         if hierarchical_ltc_phi > 0.3 {
             let hltc_boost = (hierarchical_ltc_phi - 0.3).min(0.2) * 0.05;
             self.prediction_confidence =
-                (self.prediction_confidence + hltc_boost as f32).clamp(0.0, 1.0);
+                (self.prediction_confidence + hltc_boost).clamp(0.0, 1.0);
         }
 
         // ═══════════════════════════════════════════════════════════════════════
